requirements.txt
asgiref==3.9.1
asttokens==3.0.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
crispy-tailwind==1.0.3
cryptography==45.0.5
decorator==5.2.1
Django==5.2.4
django-allauth==65.10.0
django-crispy-forms==2.4
django-environ==0.12.0
exceptiongroup==1.3.0
executing==2.2.0
idna==3.10
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.1.7
oauthlib==3.3.1
parso==0.8.4
pillow==11.3.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pymemcache==4.0.0
requests==2.32.4
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
